fix: errors

+++ raiden/tests/all/get.rs
@@ -302,7 +302,7 @@ mod tests {
                 name: "ap-northeast-1".into(),
             });
 
-            let res = client.get("id1", 2003 as usize).run().await;
+            let res = client.get("id1", 2003_usize).run().await;
             assert_eq!(
                 res.unwrap(),
                 get::GetOutput {
